Removing single quotes from logfile on WIN OS
When the start command is run on Windows, the single quotes appear as part of the logfile path or name.

We discovered this when we tried to set the log file argument to `"logfile" => "browserstack.log"` and found that the single quotes appeared as part of the file name.

Trying to set a log file to a different folder path (i.e. `"logfile" => "logs/browserstack.log"`) would cause single quotes to appear in the path itself (which caused BrowserStackLocal.exe to fail all together).

Removing the single quotes from the command string on windows resolves the issue.